Attempting to get the progress meter to work (may or may not go in an infinite loop)

+++ downloader-forwarder/source/main.cpp
@@ -18,6 +18,7 @@
 
 #include <string>
 #include <filesystem>
+#include <iostream>
 #include <switch.h>
 
 // The goal of the downloader-forwarder is to delete an older version of the app when the user wants to update it.
@@ -26,22 +27,82 @@
 
 #define PATH "/switch/Downloader/"
 #define FULL_PATH "/switch/Downloader/Downloader.nro"
+#define POSSIBLE_PATH "/switch/Downloader.nro"
 #define DOWNLOAD_CONFIG_PATH "/config/Downloader/Downloader.nro"
 #define CONFIG_PATH "/config/Downloader/"
 
+void moveToFullPath(const std::string& path)
+{
+    std::filesystem::path currentNroPath{path};
+
+    std::filesystem::copy_file(currentNroPath, std::string(FULL_PATH));
+    std::filesystem::remove(currentNroPath);
+
+    if (!std::filesystem::exists(currentNroPath.parent_path()))
+        std::filesystem::create_directories(currentNroPath.parent_path());
+}
+
+void showText(const std::string& msg) // Shows text (i.e. an error message if something went wrong)
+{
+    // Init the app with the text console
+    consoleInit(NULL);
+
+    // Set controller input layout to the standard one
+    padConfigureInput(1, HidNpadStyleSet_NpadStandard);
+
+    // Initalize the default gamepad
+    PadState pad;
+    padInitializeDefault(&pad);
+
+    std::cout << msg << '\n';
+    std::cout << "Press + to exit into the HBMenu." << '\n';
+
+    while(appletMainLoop())
+    {
+        // Update the pad to any new input
+        padUpdate(&pad);
+
+        u64 kDown = padGetButtonsDown(&pad);
+
+        if (kDown & HidNpadButton_Plus) // If the button is equal to the Plus (+) button, 
+            break; // we exit the loop.
+
+        // Otherwise, we wait for the user to exit the app.
+
+        consoleUpdate(NULL);
+    }
+
+    consoleExit(NULL);
+}
+
 int main(int argc, char *argv[])
 {
-    std::filesystem::remove(FULL_PATH);
+    if (std::filesystem::exists(POSSIBLE_PATH) || std::filesystem::exists(FULL_PATH))
+    {
+        std::string currentPath = "";
 
-    if (!std::filesystem::exists(PATH))
-        std::filesystem::create_directory(PATH);
+        if (std::filesystem::exists(POSSIBLE_PATH))
+            currentPath = POSSIBLE_PATH;
 
-    std::filesystem::copy_file(DOWNLOAD_CONFIG_PATH, FULL_PATH);
-    std::filesystem::remove(DOWNLOAD_CONFIG_PATH);
+        else if (std::filesystem::exists(FULL_PATH))
+            currentPath = FULL_PATH;
 
-    if (!std::filesystem::exists(CONFIG_PATH))
-        std::filesystem::create_directories(CONFIG_PATH);
+        std::filesystem::remove(currentPath);
+
+        if (!std::filesystem::exists(PATH))
+            std::filesystem::create_directory(PATH);
+
+        std::filesystem::copy_file(DOWNLOAD_CONFIG_PATH, FULL_PATH);
+        std::filesystem::remove(DOWNLOAD_CONFIG_PATH);
+
+        if (!std::filesystem::exists(CONFIG_PATH))
+            std::filesystem::create_directories(CONFIG_PATH);
+        
+        envSetNextLoad(FULL_PATH, ("\"" + std::string(FULL_PATH) + "\"").c_str());
+        return EXIT_SUCCESS;
+    }
+    else
+        showText("Error: Downloader NRO file does not exist in perfered path (" + std::string(FULL_PATH) + ") or possible path (" + std::string(POSSIBLE_PATH) + ").");
     
-    envSetNextLoad(FULL_PATH, ("\"" + std::string(FULL_PATH) + "\"").c_str());
-    return EXIT_SUCCESS;
+    return EXIT_FAILURE;
 }
